services/auth/pyproject.toml
[tool.poetry]
name = "auth"
version = "0.1.0"
description = ""
authors = ["serafinovsky <m.serafinovsky@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
orjson = "^3.10.12"
asyncpg = "^0.30.0"
alembic = "^1.14.0"
psycopg2-binary = "^2.9.10"
sqlalchemy = "^2.0.36"
uvicorn = "^0.34.0"
fastapi = "^0.115.6"
pydantic = "^2.10.4"
pydantic-settings = "^2.7.0"
redis = "^5.2.1"
httpx = "^0.28.1"
pyjwt = "^2.10.1"
aiofiles = "^24.1.0"
taskiq = "^0.11.10"
taskiq-redis = "^1.0.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
mypy = "^1.14.0"
ruff = "^0.9.9"
ipython = "^9.0.0"
pydantic = "^2.10.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
warn_return_any = true
warn_unused_configs = true
plugins = ["pydantic.mypy"]

[tool.pydantic-mypy]
init_typed = true

[tool.ruff]
target-version = "py313"
line-length = 100
indent-width = 4

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
